contract definition fix
contract definition fix
