+++ amorphie.contract.zeebe/Services/DocumentGroupDefinitionService.cs
@@ -7,7 +7,6 @@ using amorphie.contract.core.Entity.Document;
 using amorphie.contract.core.Entity.Document.DocumentTypes;
 using amorphie.contract.data.Contexts;
 using amorphie.contract.zeebe.Model.DocumentDefinitionDataModel;
-using amorphie.contract.zeebe.Services.Interfaces;
 using Newtonsoft.Json;
 using Microsoft.EntityFrameworkCore;
 using amorphie.contract.core.Entity.Document.DocumentGroups;
@@ -15,6 +14,10 @@ using amorphie.contract.zeebe.Model.DocumentGroupDataModel;
 
 namespace amorphie.contract.zeebe.Services
 {
+    public interface IDocumentGroupDefinitionService
+    {
+        Task<DocumentGroup> DataModelToDocumentGroupDefinition(dynamic documentDefinitionGroupDataDynamic, Guid id);
+    }
     public class DocumentGroupDefinitionService : IDocumentGroupDefinitionService
     {
         ProjectDbContext _dbContext;
