--- amorphie.contract.zeebe/Services/Interfaces/IDocumentGroupDefinitionService.cs
@@ -1,16 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using amorphie.contract.core.Entity.Common;
-using amorphie.contract.core.Entity.Document;
-using amorphie.contract.core.Entity.Document.DocumentGroups;
-using amorphie.contract.zeebe.Model.DocumentDefinitionDataModel;
-
-namespace amorphie.contract.zeebe.Services.Interfaces
-{
-    public interface IDocumentGroupDefinitionService
-    {
-        Task<DocumentGroup> DataModelToDocumentGroupDefinition(dynamic documentDefinitionGroupDataDynamic, Guid id);
-    }
-}
